Only install font if not present

+++ install/linux.sh
@@ -6,7 +6,9 @@ ln -sfv "$DOTFILES_DIR/rofi/config" ~/.config/rofi/
 . "$DOTFILES_DIR/install/nvim.sh"
 . "$DOTFILES_DIR/install/apt.sh"
 
-wget https://github.com/epk/SF-Mono-Nerd-Font/archive/refs/heads/master.zip
-7z x master.zip
-cd SF-Mono-Nerd-Font-master && mv *.otf ~/.local/share/fonts/
-cd .. && rm -rf master.zip SF-Mono-Nerd-Font-master
+if [[ ! -e "~/.local/share/fonts/SFMono Regular Nerd Font Complete.otf" ]]; then
+    wget https://github.com/epk/SF-Mono-Nerd-Font/archive/refs/heads/master.zip
+    7z x master.zip
+    cd SF-Mono-Nerd-Font-master && mv *.otf ~/.local/share/fonts/
+    cd .. && rm -rf master.zip SF-Mono-Nerd-Font-master
+fi
