+++ install/osx.sh
@@ -33,7 +33,9 @@ if [ "$LARAVEL" = "y" ]; then
 	. "$DOTFILES_DIR/install/laravel-installer.sh"
 fi
 
-wget https://github.com/epk/SF-Mono-Nerd-Font/archive/refs/heads/master.zip
-7z x master.zip
-cd SF-Mono-Nerd-Font-master && mv *.otf ~/Library/Fonts/
-cd .. && rm -rf master.zip SF-Mono-Nerd-Font-master
+if [[ ! -e "~/Library/Fonts/SFMono Regular Nerd Font Complete.otf" ]]; then
+    wget https://github.com/epk/SF-Mono-Nerd-Font/archive/refs/heads/master.zip
+    unzip -a master.zip
+    cd SF-Mono-Nerd-Font-master && mv *.otf ~/Library/Fonts/
+    cd .. && rm -rf master.zip SF-Mono-Nerd-Font-master
+fi
